Add a wpt command for updating expectation metadata.
The |wpt update-expectations| command takes a list of raw log files,
and updates the expectation data located in the directory supplied
thorugh the --metadata keyword argument. This metadata is used by
subsequent test runs to determine the expected test results.

By default the "product" property is used as a possible conditional on
the data; this means that if multiple logs are passed in
simultaneously then different results will be distinguished by "if
product==foo" conditions in the metadata.