HTMLMediaElement: set `muted_` when cloning element with muted attribute.
This CL works around the fact that cloning does not call the parser
callbacks and the HTML specifications require the attribute to only
have an effect when the element was created. There is unfortunately
no notification that the element creation from `cloneNode()` happens
so we instead rely on a callback meant for something slightly different.

Bug: 811743
Change-Id: Ib7ac55088b9eaad23f9bce0cfc8366fec6fc5b0d